==> 1-data-loading.py <==
import pandas as pd
import os
import numpy as np
import matplotlib.pyplot as plt

"""
Data loading script for Pagan data. This takes the individual CSV files outputted by Pagan and combines them into a single CSV file for each annotation type (visual, audio, engagement).
Has the option to filter out duplicate sessions (i.e. if the same participant ID annotated the task multiple times, only the latest session is kept).
At this point the raw traces only contain changes in the annotation, not the time-windowed annotation values. This is done in the next step (2-data-preprocessing.py).
"""

def latest_session(participant_data):
    all_sessions = participant_data['SessionID'].unique()
    timesteps = []
    for i in all_sessions:
        session_trace = participant_data[participant_data['SessionID'] == i]
        timesteps.append(session_trace['Timestamp'].to_numpy()[0])
    return all_sessions[np.argmax(timesteps)]


def filter_duplicate_sessions(df, engagementSession=False):
    sessions = df['PaganSession'].unique()
    latest_sessions_df = pd.DataFrame()  

    plotting = False

    for session in sessions:
        session_df = df[df['PaganSession'] == session]
        participants = session_df['Participant'].unique()
        for participant in participants:
            participant_df = session_df[session_df['Participant'] == participant]
            if engagementSession:
                games = participant_df['DatabaseName'].unique()
                for game in games:
                    game_df = participant_df[participant_df['DatabaseName'] == game]
                    latest_session_df = game_df[game_df['SessionID'] == latest_session(game_df)]

                    if not plotting and session == 'Session-1' and latest_session_df['Group'].values[0] == "Expert" and "_27" in game:
                        plt.scatter(latest_session_df['VideoTime'].values[:-1], latest_session_df['Value'].values[:-1])
                        plt.xlabel("Time (s)")
                        plt.xticks([np.min(latest_session_df['VideoTime'].values[:-1]), np.max(latest_session_df['VideoTime'].values)], ['0', '60'])
                        plt.yticks([np.min(latest_session_df['Value'].values[:-1]), np.max(latest_session_df['Value'].values)])

                        plt.ylabel("Value")
                        plt.show()

                    latest_sessions_df = pd.concat([latest_sessions_df, latest_session_df], ignore_index=True)
            else:
                latest_session_df = participant_df[participant_df['SessionID'] == latest_session(participant_df)]
                latest_sessions_df = pd.concat([latest_sessions_df, latest_session_df], ignore_index=True)
    return latest_sessions_df


def execute(FILTER_DUPLICATES=True):

    engagement_df_list = []
    brightness_df_list = []
    pitch_df_list = []

    for filename in os.listdir("./Pagan_CSVs/"):
        if filename.endswith('.csv'):
            session_name = filename.split("_")[0]
            group_name = session_name.split("-")[-1]
            session_name = session_name.removesuffix(f'-{group_name}')
            df = pd.read_csv(f"./Pagan_CSVs/{filename}").reset_index(drop=True)
            df['PaganSession'] = session_name 
            df['Group'] = group_name
            if 'Engagement' in filename:
                engagement_df_list.append(df)
            elif 'Green-Brightness' in filename:
                brightness_df_list.append(df)
            elif 'Sound-Pitch' in filename:
                pitch_df_list.append(df)

    engagement_df = pd.concat(engagement_df_list, ignore_index=True)
    brightness_df = pd.concat(brightness_df_list, ignore_index=True)
    pitch_df = pd.concat(pitch_df_list, ignore_index=True)

    if FILTER_DUPLICATES:
        green_brightness_df = filter_duplicate_sessions(brightness_df)
        sound_pitch_df = filter_duplicate_sessions(pitch_df)
        engagement_df = filter_duplicate_sessions(engagement_df, True)

    engagement_df.to_csv("./Processed Data/Raw_Engagement_Logs.csv")
    green_brightness_df.to_csv("./Processed Data/Raw_Visual_Logs.csv")
    sound_pitch_df.to_csv("./Processed Data/Raw_Audio_Logs.csv")
